Fix all findbugs warnings
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/commons/trunk@1681108 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/xmlgraphics/image/writer/internal/TIFFImageWriter.java
@@ -68,8 +68,8 @@ public class TIFFImageWriter extends AbstractImageWriter {
                 int denom;
 
                 if (ResolutionUnit.INCH == params.getResolutionUnit()) {
-                    numPixX = params.getXResolution().intValue();
-                    numPixY = params.getYResolution().intValue();
+                    numPixX = params.getXResolution();
+                    numPixY = params.getYResolution();
                     denom = 1;
                 } else {
                     // Set target resolution
